:recycle: refactor: refactor code
