+++ src/components/Items.js
@@ -1,20 +1,24 @@
 import { useState, useEffect } from 'react';
 import Container from './Container';
+import List from './List';
 
-const Item = () => {
+// 할 일들을 관리(추가/삭제/토글)하는 컴포넌트
+const Items = () => {
   // items: 전체 할 일들을 저장
   const [items, setItems] = useState([]);
 
+  // local storage에 저장된 할 일 불러오기
   useEffect(() => {
     const savedItems = localStorage.getItem('todoItems');
     savedItems && setItems(JSON.parse(savedItems));
   }, []);
 
+  // local storage에 items 저장하기
   useEffect(() => {
     localStorage.setItem('todoItems', JSON.stringify(items));
   }, [items]);
 
-  // todo: 입력한 할 일
+  // inputText: 입력한 할 일
   const [inputText, setInputText] = useState('');
   const [id, setId] = useState(0);
 
@@ -24,13 +28,16 @@ const Item = () => {
     isDone: false,
   };
 
+  // 인풋 필드에 입력한 값으로 inputText를 변경
   const handleChange = (e) => {
     setInputText(e.target.value);
   };
 
+  // items 배열에 입력한 할 일을 추가
   const addNewTodo = (e) => {
     e.preventDefault();
 
+    // 중복 검사
     const index = items.findIndex((todoObj) => todoObj.text === inputText);
 
     if (inputText && index === -1) {
@@ -41,10 +48,12 @@ const Item = () => {
     setId(id + 1);
   };
 
+  // items 배열에서 할 일 삭제
   const deleteTodo = (id) => {
     setItems(items.filter((todoObj) => todoObj.id !== id));
   };
 
+  // 할 일의 isDone을 반전
   const toggleTodo = (id) => {
     setItems(
       items.map((todo) =>
@@ -58,12 +67,11 @@ const Item = () => {
       todo={inputText}
       handleChange={handleChange}
       addNewTodo={addNewTodo}
-      deleteTodo={deleteTodo}
-      items={items}
-      setItems={setItems}
-      toggleTodo={toggleTodo}
+      List={
+        <List items={items} deleteTodo={deleteTodo} toggleTodo={toggleTodo} />
+      }
     />
   );
 };
 
-export default Item;
+export default Items;
